Extract 'Linear' interface
